+++ src/main/java/vesterosse/library/Book.java
@@ -1,10 +1,11 @@
 package vesterosse.library;
 
 import lombok.AllArgsConstructor;
+import lombok.Getter;
 
-import java.util.Map;
 import java.util.Objects;
 
+@Getter
 @AllArgsConstructor
 public class Book {
     private String title;
@@ -18,8 +19,12 @@ public class Book {
 
     @Override
     public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
+        if (this == o) {
+            return true;
+        }
+        if (o == null || getClass() != o.getClass()) {
+            return false;
+        }
         Book book = (Book) o;
         return year == book.year && Objects.equals(title, book.title) && Objects.equals(author, book.author);
     }
@@ -29,26 +34,5 @@ public class Book {
         return "Название: '" + title + '\'' +
                 ", Автор: '" + author + '\'' +
                 ", Год: " + year;
-    }
-
-    public static void addBook(Map<Book, String> map, Book book, String place) {
-        map.put(book, place);
-        System.out.println("Книга '" + book.title + "' добавлена в библиотеку!" + "\n");
-    }
-
-    public static void removeBook(Map<Book, String> map, String title, String author, int year) {
-        map.remove(new Book(title, author, year));
-        System.out.println("Вы удалили книгу '" + title + "' из библиотеки!" + "\n");
-    }
-
-    public static void searchBook(Map<Book, String> map, String title, String author, int year) {
-        System.out.println("Ваша книга найдена! Она находится на полке № " + map.get(new Book(title, author, year)) + "\n");
-    }
-
-    public static void printBook(Map<Book, String> map) {
-        System.out.println("Список всех книг в библиотеке:");
-        for (Map.Entry<Book, String> entry : map.entrySet()) {
-            System.out.println(entry.getKey() + ", Полка № " + entry.getValue());
-        }
     }
 }
